+++ CHANGELOG.md
@@ -104,6 +104,7 @@
 - PR #2022 Facilitating get_dummies to be a high level api too
 - PR #2050 Namespace IO readers and add back free-form `read_xxx` functions
 - PR #2104 Add a functional ``sort=`` keyword argument to groupby
+- PR #2108 Add `find_and_replace` for StringColumn for replacing single values
 
 ## Bug Fixes